+++ CHANGELOG.rst
@@ -1,3 +1,3 @@
 Unreleased
 ----------
-
+* new method: `Introduction.introduction`